+++ resolwe/api_urls.py
@@ -26,9 +26,6 @@ api_router.register(r"entity", EntityViewSet)
 api_router.register(r"relation", RelationViewSet)
 api_router.register(r"descriptorschema", DescriptorSchemaViewSet)
 api_router.register(r"storage", StorageViewSet)
-api_router.register(r"upload/config", UploadConfig, basename="upload_config")
-# TODO: The view below is deprecated and must be removed on 1. 1. 2025.
-api_router.register(r"upload_config", UploadConfig, basename="upload_config_deprecated")
 api_router.register(r"task", BackgroundTaksViewSet, basename="backgroundtask")
 api_router.register(r"annotation_field", AnnotationFieldViewSet)
 api_router.register(r"annotation_preset", AnnotationPresetViewSet)
@@ -38,8 +35,9 @@ api_router.register(r"annotation_value", AnnotationValueViewSet)
 urlpatterns = [
     path("api/", include((api_router.urls, "resolwe-api"), namespace="resolwe-api")),
     path(
-        "api/upload/credentials",
-        UploadCredentials.as_view(),
-        name="upload_credentials",
+        "api/upload/credentials", UploadCredentials.as_view(), name="upload_credentials"
     ),
+    path("api/upload/config", UploadConfig.as_view(), name="upload_config"),
+    # TODO: The view below is deprecated and must be removed on 1. 1. 2025.
+    path("api/upload_config", UploadConfig.as_view(), name="upload_config_deprecated"),
 ]